+++ src/util/error.handler.ts
@@ -1,26 +1,8 @@
 import { Router, Request, Response, NextFunction } from "express";
 import { fail } from "./response";
-
-export default class CustomHttpError extends Error {
-    httpStatusCode;
-    timestamp;
-   
-    constructor(httpStatusCode: any, message: string) {
-      if (message) {
-        super(message);
-      } else {
-        super("A generic error occurred!");
-      }
-   
-      // initializing the class properties
-      this.httpStatusCode = httpStatusCode;
-      this.timestamp = new Date().toISOString();
-   
-      // attaching a call stack to the current class,
-      // preventing the constructor call to appear in the stack trace
-      // Error.captureStackTrace(this, this.constructor);
-    }
-  }
+import ValidationError from "./validation.error";
+import CustomHttpError from "./custom.error";
+ 
 
 /**
  * Init Express error handler
@@ -32,11 +14,14 @@ export const errorHandler = (err: Error, req: Request, res: Response, next: Next
     // default HTTP status code and error message
     let httpStatusCode = 400;
     let message = "Error Occurred";
+
+    console.log(err instanceof CustomHttpError);
    
     // if the error is a custom defined error
     if (err instanceof CustomHttpError) {
-      httpStatusCode = err.httpStatusCode;
+      console.log("Custom Error")
       message = err.message;
+      httpStatusCode = err.httpStatusCode;
     } 
     else 
     {
@@ -58,7 +43,7 @@ export const errorHandler = (err: Error, req: Request, res: Response, next: Next
     }
    
     // logg the error
-    console.error(err);
+    //console.error(err);
     // other custom behaviors...
    
     // return the standard error response
