+++ src/util/validation.error.ts
@@ -1,0 +1,5 @@
+import CustomHttpError from "./custom.error";
+
+export default class ValidationError extends CustomHttpError {
+    
+}
